test(master-data): split 1424-line test_repositories.py by class group (#109, A-8)
Splits services/master-data/tests/unit/test_repositories.py (1424
lines, 9 test classes) into three focused files. Each new file is
under ~705 lines.

  test_item_master_repositories.py   (704 lines) — Item-related:
    Category, ItemCommon, ItemStore, ItemBook.

  test_settings_payment_repositories.py (498 lines) — Settings-side:
    Settings, Payment, Promotion.

  test_staff_tax_repositories.py     (263 lines) — Staff, Tax.

Verified: pipenv run pytest -m unit → 380 passed (same as before).

Refs #109, follows up #113.
